renaming for future updates where it will support uio and custom drivers also
